major update
Moved lots of new functions into the repository.  They may not all work yet with the publicly available database and not all codes are well commented.  If you would like comments or a specific code fixed, send me a message and I should be able to get it working in quickly.